refactoring project structure add dto/impl folder. DTO's replace in impl folder. Add getQuestionDTO to QuestionDTO
