+++ src/main/java/io/khasang/jrepetitor/controller/QuestionController.java
@@ -1,6 +1,5 @@
 package io.khasang.jrepetitor.controller;
 
-import io.khasang.jrepetitor.dto.QuestionDTO;
 import io.khasang.jrepetitor.dto.QuestionDTOInterface;
 import io.khasang.jrepetitor.entity.Question;
 import io.khasang.jrepetitor.service.QuestionService;
