+++ src/main/java/io/khasang/jrepetitor/controller/QuizController.java
@@ -1,8 +1,7 @@
 package io.khasang.jrepetitor.controller;
 
-import io.khasang.jrepetitor.dto.QuizDTO;
 import io.khasang.jrepetitor.dto.QuizDTOInterface;
-import io.khasang.jrepetitor.dto.QuizPreviewDTO;
+import io.khasang.jrepetitor.dto.impl.QuizPreviewDTO;
 import io.khasang.jrepetitor.entity.Quiz;
 import io.khasang.jrepetitor.service.QuizService;
 import org.springframework.beans.factory.annotation.Autowired;
