+++ src/main/java/io/khasang/jrepetitor/controller/UserController.java
@@ -1,6 +1,6 @@
 package io.khasang.jrepetitor.controller;
 
-import io.khasang.jrepetitor.dto.UserDTO;
+import io.khasang.jrepetitor.dto.impl.UserDTO;
 import io.khasang.jrepetitor.entity.Profile;
 import io.khasang.jrepetitor.entity.User;
 import io.khasang.jrepetitor.service.UserService;
