+++ src/main/java/io/khasang/jrepetitor/dto/impl/QuizDTO.java
@@ -1,5 +1,7 @@
-package io.khasang.jrepetitor.dto;
+package io.khasang.jrepetitor.dto.impl;
 
+import io.khasang.jrepetitor.dto.QuestionDTOInterface;
+import io.khasang.jrepetitor.dto.QuizDTOInterface;
 import io.khasang.jrepetitor.entity.Item;
 import io.khasang.jrepetitor.entity.Question;
 import io.khasang.jrepetitor.entity.Quiz;
