+++ src/main/java/io/khasang/jrepetitor/dto/impl/QuizPreviewDTO.java
@@ -1,4 +1,4 @@
-package io.khasang.jrepetitor.dto;
+package io.khasang.jrepetitor.dto.impl;
 
 import io.khasang.jrepetitor.entity.Quiz;
