+++ src/main/java/io/khasang/jrepetitor/service/impl/GroupServiceImpl.java
@@ -1,7 +1,7 @@
 package io.khasang.jrepetitor.service.impl;
 
 import io.khasang.jrepetitor.dao.GroupDao;
-import io.khasang.jrepetitor.dto.GroupDTO;
+import io.khasang.jrepetitor.dto.impl.GroupDTO;
 import io.khasang.jrepetitor.entity.Group;
 import io.khasang.jrepetitor.service.GroupService;
 import org.springframework.beans.factory.annotation.Autowired;
