+++ src/main/java/io/khasang/jrepetitor/service/impl/ItemServiceImpl.java
@@ -1,7 +1,7 @@
 package io.khasang.jrepetitor.service.impl;
 
 import io.khasang.jrepetitor.dao.ItemDao;
-import io.khasang.jrepetitor.dto.ItemDTO;
+import io.khasang.jrepetitor.dto.impl.ItemDTO;
 import io.khasang.jrepetitor.entity.Item;
 import io.khasang.jrepetitor.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
